developer-environment/requirements.txt
-e .
aws-cdk.aws-ec2
aws-cdk.aws-rds
aws-cdk.core
